+++ src/collective/frontpage/config.py
@@ -1,0 +1,96 @@
+# -*- coding: utf-8 -*-
+
+DUMMY_TITLES = [
+    u'Quo in',
+    u'Aspernatur zzril',
+    u'Deleniti accusam',
+    u'Repellat kasd',
+    u'Consectetur vero',
+    u'Veritatis sanctus',
+    u'Architecto numquam',
+    u'Pariatur gubergren',
+    u'Magnam dolores te',
+    u'Maxime blandit',
+    u'Lorem ipsum',
+]
+
+DUMMY_SUBTITLES = [
+    u'Iriure dolor in hendrerit.',
+    u'Vero eros et accumsan.',
+    u'Praesent luptatum zzril.',
+    u'Autem vel eum iriure.',
+    u'Stet clita kasd.',
+    u'Vero eos et accusam.',
+    u'Dignissim qui blandit.',
+    u'Justo duo dolores et.',
+    u'Consetetur sadipscing elitr.',
+    u'Clita kasd gubergren.',
+    u'Te feugait nulla.',
+    u'Nulla facilisis at vero.',
+    u'Takimata sanctus est.',
+]
+
+DUMMY_DESCRIPTIONS = [
+    u'Duis autem vel eum iriure dolor in hendrerit in vulputate velit '
+    'esse molestie consequat.',
+    u'Feugiat nulla facilisis at vero eros et accumsan et iusto odio '
+    'dignissim qui blandit praesent luptatum zzril delenit augue duis '
+    'dolore te feugait nulla facilisi. Lorem ipsum dolor sit',
+    u'In hendrerit in vulputate velit esse molestie consequat, vel illum '
+    'dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto '
+    'odio dignissim qui blandit praesent.',
+    u'At vero eos et accusam et justo duo dolores et ea rebum. '
+    'Stet clita kasd gubergren.',
+    u'Takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor '
+    'sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor '
+    'invidunt ut labore et dolore.',
+    u'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam '
+    'nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam '
+    'erat, sed diam voluptua.',
+    u'Stet clita kasd gubergren, no sea takimata sanctus est '
+    'Lorem ipsum dolor sit amet.',
+    u'Magna aliquyam erat, sed diam voluptua. At vero eos et accusam '
+    'et justo duo dolores et ea rebum.',
+    u'Zzril delenit augue duis dolore te feugait nulla facilisi. Lorem '
+    'ipsum dolor sit amet.',
+    u'Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper '
+    'suscipit lobortis nisl ut aliquip ex ea commodo consequat.',
+    u'Nam liber tempor cum soluta nobis eleifend option congue nihil '
+    'imperdiet doming id quod mazim placerat facer possim assum.',
+    u'Sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna '
+    'aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud '
+    'exerci tation ullamcorper suscipit lobortis nisl ut aliquip.',
+    u'Vel eum iriure dolor in hendrerit in vulputate velit esse '
+    'molestie consequat, vel illum dolore eu feugiat.',
+]
+
+DUMMY_TEXT = [
+    u'<p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed '
+    'do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim '
+    'ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut '
+    'aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit '
+    'in voluptate velit esse cillum dolore eu fugiat nulla pariatur.'
+    '</p><p>Excepteur sint occaecat cupidatat non proident, sunt in culpa qui '
+    'officia deserunt mollit anim id est laborum. Lorem ipsum dolor sit '
+    'amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt '
+    'ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud '
+    'exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. '
+    'Duis aute irure dolor in reprehenderit in voluptate velit esse cillum '
+    'dolore eu fugiat nulla pariatur.</p>',
+    u'<p>Consectetur adipisicing elit, sed do eiusmod tempor incididunt '
+    'ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis '
+    'nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo '
+    'consequat. Duis aute irure dolor in reprehenderit in voluptate velit '
+    'esse cillum dolore eu fugiat nulla pariatur.</p>',
+    u'<p>Sed do eiusmod tempor incididunt ut labore et dolore magna '
+    'aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco '
+    'laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor '
+    'in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla '
+    'pariatur.</p><p>Excepteur sint occaecat cupidatat non proident, sunt in '
+    'culpa qui officia deserunt mollit anim id est laborum. Lorem ipsum '
+    'dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor '
+    'incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, '
+    'quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo '
+    'consequat. Duis aute irure dolor in reprehenderit in voluptate velit '
+    'esse cillum dolore eu fugiat nulla pariatur.</p>'
+]